Move checkers in pkg/types/clusterconfig
